Updates import for functions moved to date-range.js.
