+++ assets/js/modules/analytics/dashboard/dashboard-widget-top-acquisition-sources-table.js
@@ -31,10 +31,10 @@ import { Component, Fragment } from '@wordpress/element';
  * Internal dependencies
  */
 import {
-	getCurrentDateRange,
 	getTimeInSeconds,
 	numberFormat,
 } from '../../../util';
+import { getCurrentDateRange } from '../../../util/date-range';
 import withData from '../../../components/higherorder/withdata';
 import { TYPE_MODULES } from '../../../components/data';
 import { getDataTableFromData, TableOverflowContainer } from '../../../components/data-table';
