+++ assets/js/modules/search-console/dashboard/dashboard-widget.js
@@ -40,7 +40,7 @@ import Alert from '../../../components/alert';
 import ProgressBar from '../../../components/progress-bar';
 import getNoDataComponent from '../../../components/notifications/nodata';
 import getDataErrorComponent from '../../../components/notifications/data-error';
-import { getCurrentDateRange } from '../../../util';
+import { getCurrentDateRange } from '../../../util/date-range';
 import HelpLink from '../../../components/help-link';
 
 class GoogleSitekitSearchConsoleDashboardWidget extends Component {
