+++ assets/js/util/index.js
@@ -53,7 +53,7 @@ export * from './stringify';
 export * from './standalone';
 export * from './storage';
 export * from './i18n';
-export * from './date-range';
+//export * from './date-range';
 export * from './helpers';
 
 /**
